requirements.txt
absl-py==2.3.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albumentations==1.1.0
altair==5.5.0
annotated-types==0.7.0
annoy==1.17.3
anyio==4.10.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.2
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.4.0
distro==1.9.0
Django==5.2.3
djangorestframework==3.16.0
dnspython==2.7.0
einops==0.8.1
evaluate==0.4.3
eventlet==0.40.0
exceptiongroup==1.3.0
executing==2.2.0
facets-overview==1.1.1
faiss-cpu==1.12.0
filelock==3.18.0
Flask==3.1.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.11.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kagglehub==0.3.12
keras==2.15.0
kiwisolver==1.4.7
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langsmith==0.4.21
lazy_loader==0.4
Levenshtein==0.27.1
libclang==18.1.1
lightning-utilities==0.15.1
loguru==0.7.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
MedPy==0.5.2
ml-dtypes==0.2.0
ml_collections==1.1.0
mmcv-full==1.7.2
monai==1.4.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.3.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nltk==3.9.1
numpy==2.3.2
oauthlib==3.3.1
openai==1.102.0
opencv-python==4.5.5.64
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.3
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
ptflops==0.7.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
python-youtube==0.9.8
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qudida==0.0.4
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
segmentation-mask-overlay==0.3.4
sentencepiece==0.2.0
simple-websocket==1.1.0
simpleitk==2.5.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.49.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
tensorflow-metal==1.2.0
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.11.0
timm==0.6.12
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.8.0
torchprofile==0.0.4
torchsummary==1.5.1
torchsummaryX==1.3.0
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.21.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
warmup_scheduler==0.3
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
youtube-transcript-api==1.2.2
zipp==3.23.0
zstandard==0.24.0
